requirements.txt
json
tensorflow-gpu